# backend/drone-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления дронами доставки",
        "title": "SkyPost Drone Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "skr1ms13666@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
        "/api/drones/{drone_id}/command": {
            "post": {
                "description": "Отправляет команду управления конкретному дрону",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drones"
                ],
                "summary": "Отправить команду дрону",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID дрона",
                        "name": "drone_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Команда для дрона",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_request.SendCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.CommandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Error"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Проверка состояния сервиса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitoring"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Health"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Возвращает текущий статус первого зарегистрированного дрона",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drones"
                ],
                "summary": "Получить статус дрона",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.DroneStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Error"
                        }
                    }
                }
            }
        },
        "/ws/admin": {
            "get": {
                "description": "Устанавливает WebSocket соединение для мониторинга дронов администратором\nАдминистратор получает обновления статуса всех дронов и доставок в реальном времени",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket подключение администратора",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID администратора",
                        "name": "admin_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ws/drone": {
            "get": {
                "description": "Устанавливает WebSocket соединение для дрона. Первое сообщение должно быть регистрацией с ip_address\nПоддерживаемые типы сообщений: heartbeat, status_update, delivery_update, video_frame, arrived_at_destination, cargo_dropped",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket подключение дрона",
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ws/drone/{drone_id}": {
            "get": {
                "description": "Устанавливает WebSocket соединение для дрона. Первое сообщение должно быть регистрацией с ip_address\nПоддерживаемые типы сообщений: heartbeat, status_update, delivery_update, video_frame, arrived_at_destination, cargo_dropped",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket подключение дрона",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID дрона (опционально)",
                        "name": "drone_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ws/drone/{drone_id}/video": {
            "get": {
                "description": "Устанавливает WebSocket соединение для получения видео потока от конкретного дрона\nАдминистратор подключается к этому эндпоинту для просмотра видео в реальном времени",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket для видео потока дрона",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID дрона",
                        "name": "drone_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "drone_id required",
                        "schema": {
                            "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_request.SendCommand": {
            "type": "object",
            "required": [
                "command"
            ],
            "properties": {
                "command": {
                    "type": "string",
                    "example": "start_delivery"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.CommandResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Command successfully queued for delivery"
                },
                "status": {
                    "type": "string",
                    "example": "command_sent"
                }
            }
        },
        "github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.DroneStatus": {
            "type": "object",
            "properties": {
                "battery_level": {
                    "type": "number",
                    "example": 85.5
                },
                "current_delivery_id": {
                    "type": "string",
                    "example": "delivery-123"
                },
                "drone_id": {
                    "type": "string",
                    "example": "drone-001"
                },
                "error_message": {
                    "type": "string",
                    "example": ""
                },
                "position": {
                    "$ref": "#/definitions/github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Position"
                },
                "speed": {
                    "type": "number",
                    "example": 15.3
                },
                "status": {
                    "type": "string",
                    "example": "idle"
                }
            }
        },
        "github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Health": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string",
                    "example": "drone-service"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                }
            }
        },
        "github_com_skr1ms_SkyPostDelivery_drone-service_internal_controller_http_v1_response.Position": {
            "type": "object",
            "properties": {
                "altitude": {
                    "type": "number",
                    "example": 150
                },
                "latitude": {
                    "type": "number",
                    "example": 55.751244
                },
                "longitude": {
                    "type": "number",
                    "example": 37.618423
                }
            }
        }
    }
}